requirements.txt
streamlit==1.25.0
requests==2.31.0
Pillow==9.1.0
runpod
boto3